Announce number of results to screen readers
On opening the results `ul`, but before selecting an item,
populate the status element with the number of items in the `ul`
and with instructions to use the up and down arrow keys to navigate.

When the `ul` is closed, the status's textContent is cleared.

For localization's sake, the string inserted into the status element
is generated by a configurable function called `initialStatus`, which
receives the number of results in the `ul` as its only argument.