Various data formats and separate title/value.
New "data" configuration property is function, which is called
for each initial list item in any format to convert it to object
like this one:

{ title: "United States", value: "US" }

This object is then used everywhere in API. So instead of item text
filter(), sort(), item() and replace() functions now get an object
with separate "title" and "value" properties.

Despite this data format change, old code will continue to work
as before. This is what Suggestion() shim is for. It uses "title"
property when string is expected everywhere in the API.

The only change is that now we have "value" property and it would
be used in default implementation of replace() method directly,
so now suggestion title you see in completer and the value inserted
to the input when the item is selected can be independent.