fix: allow selection on mobile ios
On mobile ios, no mousedown event is fired, because ios keeps the focus
on the currently active input element. However, it does fire a click event.

To work around this problem, but keep the current behavior, we add a click event
handler to the existing mousedown handler, which will now only preventDefault,
which keeps blur from firing, but allows the click handler to do the actual selection.

This new behavior additionally means that clicking and holding (dragging) the mouse away
from the autocomplete list will not result in a selection, mimicking the normal
mouse behavior.

Closes #16712