Avoid script injections when rendering suggestions.
Currently (before this change), any HTML contained in the suggestion text
will be rendered by `$.create` due to the fact that `innerHTML` is being used.

This means that if there are any scripts in the suggestion text, those
scripts will be executed, which is a security risk.

For example, if the suggestion text is `<img src="fak" onerror="alert(123)">`
and the user types the text `<`, then you'll see an alert popup showing "123"
as that suggestion text gets inserted and rendered.

What this change does is to render as text (using `textContent`) the suggestion
text, while making sure that the `<mark>` elements are still rendered as HTML.